refactor: decomposing; keyboard is working! yay
