+++ src/components/App/index.jsx
@@ -1,4 +1,6 @@
 import { Keyboard } from "../Keyboard";
+import Themes from "../Themes/Themes";
+import UI from "../UI";
 
 export const App = () => {
   return (
